IGBF-1842 Update python packages using TestAppStore
When attempting to stand up App Store on new production server,
ran into problems installing mysqldb client and freeze. Logged in
to the TestAppStore host, started the virtual environment, and
ran pip freeze to get a listing of all installed python3 packages
in the virtual environment used to run the AppStore under apache.
Used this output as the new python_packages.txt file.

Note that versions of each requirement are included. This seems
a safer, more bullet-proof way to specify the different packages
and libraries required by AppStore.